Update 삼삼한 수.cs

+++ 백준/Silver/17252. 삼삼한 수/삼삼한 수.cs
@@ -15,7 +15,7 @@ namespace _17252
 
             while (n > 0)
             {
-                long remainder = n % 3;
+                int remainder = n % 3;
                 if (remainder > 1)
                 {
                     return false;
